Try to fix file descriptor exhaustion in the Docker backend. Each runner instance would creates its own Docker client.
This scales poorly. The fix used here is to have a Docker client
pool such that each thread gets its own client.

We should try removing the releases which should give us slightly
better performance. If we do that we could also sink the call to
ping in the pool so we avoid doing a ping for every runner instance
(instead we'd do it for every thread).