Change how ranking of false positives work.
Previously we would compare based on the number of false positives.
Ally (@afd) has argued that this isn't right because that assumes each
false positive is equally important which is an assumption we cannot
make.

Instead we rank based on the binary property "has false positives".

In practice for the results I'm currently analysing this has made
no difference to the ranking.