Work around bug in InvocationInfo schema.
When an InvocationInfo object is created it adds implicit fields
but it sets two implicit fields to `None` which doesn't match the
schema which means if this gets serialized and then validated against
the schema it will fail.

I've made the schema looser now to support the `None` value but the
next time we do a schema upgrade which should drop this.