Change how docker client pool works.
We now maintain a fixed pool of clients. The number of clients
is equal to the number of jobs we intend to run in parallel.

The current running thread of the caller is no longer used to identify
which resource was taken. Instead the caller must return the client
that they used.